Refactor frontend normalization helpers
