+++ contract_review_app/contract_review_app/static/panel/app/assets/__tests__/props/normalized_equivalence.spec.ts
@@ -2,6 +2,7 @@ import { describe, it, expect, beforeAll } from 'vitest';
 
 let normalizeSnippetForSearch: (typeof import('../../anchors'))['normalizeSnippetForSearch'];
 let normalizeIntakeText: (typeof import('../../normalize_intake'))['normalizeIntakeText'];
+let normalizeTextFull: (typeof import('../../normalize_full'))['normalizeTextFull'];
 let normalizeText: (typeof import('../../dedupe'))['normalizeText'];
 
 beforeAll(async () => {
@@ -9,11 +10,47 @@ beforeAll(async () => {
   normalizeSnippetForSearch = anchorsMod.normalizeSnippetForSearch;
   const intakeMod = await import('../../normalize_intake');
   normalizeIntakeText = intakeMod.normalizeIntakeText;
+  const fullMod = await import('../../normalize_full');
+  normalizeTextFull = fullMod.normalizeTextFull;
   const dedupeMod = await import('../../dedupe');
   normalizeText = dedupeMod.normalizeText;
 });
 
 describe('normalizeIntakeText', () => {
+  const CASES: Array<{ name: string; input: string; expected: string }> = [
+    {
+      name: 'NFC input remains stable',
+      input: 'Alpha - "beta\'s" clause\nline',
+      expected: 'Alpha - "beta\'s" clause\nline',
+    },
+    {
+      name: 'NFD combines and smart punctuation',
+      input: 'Cafe\u0301 — “Quote”',
+      expected: 'Café - "Quote"',
+    },
+    {
+      name: 'zero-width characters removed',
+      input: 'A\u200Bgreement\u200C',
+      expected: 'Agreement',
+    },
+    {
+      name: 'collapses whitespace and nbsp',
+      input: '\u00A0Foo\t\tBar\u202F',
+      expected: 'Foo Bar',
+    },
+    {
+      name: 'normalizes carriage returns',
+      input: 'Line1\r\nLine2',
+      expected: 'Line1\nLine2',
+    },
+  ];
+
+  it('normalizes representative variants to canonical form', () => {
+    for (const sample of CASES) {
+      expect(normalizeIntakeText(sample.input)).toBe(sample.expected);
+    }
+  });
+
   const base = 'Alpha - "beta\'s" clause';
   const replacements: Record<string, string[]> = {
     '"': ['"', '“', '”', '„', '‟', '«', '»'],
@@ -66,11 +103,35 @@ describe('normalizeIntakeText', () => {
     const sample = ' Clause\u00A0“Alpha” ';
     expect(normalizeIntakeText(sample)).toBe(normalizeIntakeText(normalizeText(sample)));
   });
+
+  it('produces the same output via full normalization helper', () => {
+    for (const sample of CASES) {
+      expect(normalizeIntakeText(sample.input)).toBe(normalizeTextFull(sample.input).text);
+    }
+  });
 });
 
 describe('normalizeSnippetForSearch', () => {
   it('delegates to intake normalization', () => {
     const sample = 'A\u200Bgreement — “Quote”\u00A0';
     expect(normalizeSnippetForSearch(sample)).toBe(normalizeIntakeText(sample));
+  });
+});
+
+describe('normalizeTextFull', () => {
+  it('returns offset map aligned with normalized text', () => {
+    const sample = ' Cafe\u0301\u200B — “Q”\r\nLine\u00A0';
+    const result = normalizeTextFull(sample);
+    expect(result.text).toBe('Café - "Q"\nLine');
+    expect(result.map).toEqual([1, 2, 3, 4, 6, 7, 8, 9, 10, 11, 12, 14, 15, 16, 17]);
+    expect(result.map).toHaveLength(result.text.length);
+  });
+
+  it('keeps map length in sync for representative samples', () => {
+    const samples = ['Cafe\u0301', 'A\u200B', ' Foo\tBar ', 'Line1\r\nLine2'];
+    for (const sample of samples) {
+      const result = normalizeTextFull(sample);
+      expect(result.map).toHaveLength(result.text.length);
+    }
   });
 });
